post like added

+++ article/apiv1/post/PostOperationResources.py
@@ -21,11 +21,18 @@ post_operation_apis = Namespace('post_operation_apis', description='Post related
 class PostOperationResource(Resource):
     """Post Delete & Update"""
 
+    @jwt_required()
     def delete(self, post_id):
         try:
             verify_jwt_in_request()
             try:
                 post = Post.get_post_by_id(post_id)
+
+                current_user_id = get_jwt_identity()
+            
+                if post.user_id != current_user_id:
+                    return {'status': False, 'message': 'Unauthorized to update this comment'}
+                
                 if post:
                     db.session.delete(post)
                     db.session.commit()
@@ -40,14 +47,16 @@ class PostOperationResource(Resource):
 
         except Exception as err:
             return {'status': False, 'message': str(err.args)}
-
+        
+    @jwt_required()
     def get(self, post_id):
         """Get a post with id"""
 
         post = Post.get_post_by_id(post_id)
         message = 'Post fetched successfully'
         return {'status': True, 'message': message, 'data': post_schema.dump(post)}
-
+    
+    @jwt_required()
     def put(self, post_id):
         """Update a post with id"""
         try:
@@ -57,6 +66,12 @@ class PostOperationResource(Resource):
 
             try:
                 post = Post.get_post_by_id(post_id)
+
+                current_user_id = get_jwt_identity()
+            
+                if post.user_id != current_user_id:
+                    return {'status': False, 'message': 'Unauthorized to update this comment'}
+                
                 if post:
                     post.title = title
                     post.content = content
@@ -72,7 +87,8 @@ class PostOperationResource(Resource):
 
         except Exception as err:
             return {'status': False, 'message': str(err.args)}
-
+        
+    @jwt_required()
     def post(self):
         """List all posts"""
         try:
